Updated the draft
Updated to 0.2. Still messy and incomplete, but will work as a
conversation point between my supervisor and myself to decide what
content the OPS should hold in the Release version. Not committing to
any detailed text just yet, as the content currently written may have to
be scrapped after this meeting if I misunderstood the project aims.