Release 1.0
Not super happy with it, but I don't know where else to go with it right
now. Foggy headed. This isn't assessed, but I will get feedback on it.
I'll look again tomorrow morning to see what else I can do, but for now
I'm submitting it so I can get help and feedback on how to improve my
writing and what is expected of me. I feel I've missed the mark a little
bit though.