File read and half GC content
Reads in files (doesn't care if they're .fa or not yet, nor does it use
any of the first line of the file yet, as this was out of scope for the
current Sprint) and then breaks it into Windows (specified in the
application, no way for an external user to change yet). The count of
the G and C is then taken, but no report or output is produced based on
this.