Moving files over to JSB
Started moving the files from the static app to the Spring Boot app.
Seems to have moved across okay. No functionality yet tied between the
two though. Need to learn how to use some JS in the produced web page,
and get the results of the processing in the Java app.