More tests and working with ORF Locations
Added more tests (TDD.. >>). Much needed tests, and used TDD to work
with the ORF Locations. There's still a lot of work that needs to go
into the ORF Locations, as right now it just does the full string. What
should actually happen is the framing (1 - 1, 2 - 2, 3 -3 and then
backwards with the matching base pair). Right now it's a naive approach,
but as a prototype... Needs something to also display with the GC in the
browser.