Some nice UI tweaks
Made the front end UI look shiny and good. Now has a logo, horay.
In the code side, just removed some printlns and added in the ORF to
quality check (code itself is still not completed for this though. Also
changed GC threshold to be based on # of standard deviations away,
rather than percentage.