More with ORF Locations
Took longer than I thought and many different iterations of how to do
this. Still needs refactoring but the TDD helped keep things in check to
just get it working. So, we now find ORFs in the first 3 frames, and it
deals with redundant codons within those frames (codons that would make
an ORF location shorter than taking another codon).
Next is the other three frames (the pairs in 'reverse' order), labelling
the frames in the result, and combining the results into a manageable
results class for use in the display.