ORF Location Canvas/View prototype
Prototype for the view of ORF Locations. Using JavaScript HTML5 canvas.
Right now it displays using some dummy data and finds when a click is on
the ORF Locations and puts up an alert. This is setup so I can format
the real data within the application itself to be consumed by the
JavaScript, and then tie it in to output this data
(Start/End/Content/Length). Will prototype this before moving it into
the WebApplication