User pasting and ORF chart
Got it so a user can paste their own sequence into the application and
see results now.
Also added the ORF Chart prototype and got it displaying real data
results, kinda. They display, but they don't fit the display yet. I need
to work on the conversion so that the chart of set length correlates to
100% of the length of the sequence, so that the start/stop locations of
the ORF locations fit at the right positions of the resulting chart.