Parameters per contig
Moved the parameters for the QA into each contig, as a hidden form that
appears when you click select to edit and then submit.
The actual layout needs work, and the boxes need a method to hide all
those previously shown, but it works so far.
Need to check that this values go through, and potentially work on
having individual values for each contig as 'suggestions' based on the
contig length for the user.