ORF Location index labels and loading button on Inspect
Clicking Inspect after the parameter settings will now change the value of the button to say 'Loading...', just to give the user some feedback that they have actually clicked the button. A dedicated loading page would be better, but is out of scope for this project.
Also, I've added labels for the current index for the first character of each line in the display of the ORF Location characters. Just makes it look a little nicer. Along with this, I cleaned up the colouring of the first and last character as there were a few instances where the colouring would be misplace/missed entirely, caused by the ordering of the conditional statements in the code.