Further improved display
Gave a scroll box to the list of ORF Locations for when there are a lot of them. Added the number of contigs in the shown list and number of discarded contigs to the contig list after input. Involved adding a new class ContigResult in order to hold the previously kept list of contigs for use in the  inspection and the new discarded variable. Small misc css changes here and there.