More view changes, sorting the list of ORF Locs by length
As title, changed the scroll box for ORF Locations again just because I didn't like how the headers were lined up. Also sorted the list of ORF Locations by longest to shorest in the model, and this comes through to the view to be seen. The clicking of the ORFs still works, as even in a different order the ids/indexes of the ORF Locations in the list are the same (as that's how they're put down.