Removed unnecessary ORF code for stop Codons based on feedback
Feedback from Amanda made me realize that the first Stop Codon is the only stop Codon and I'd made a wrong assumption. Cleaned out the code that looked for all the other Stop Codons and corrected it and my tests. Seems alright now, and matches NCBI better, but is also a little sad as that code was nice, even if wrong hah.