Code for clicking on Superframe window segments
Put in the JS needed to allow a user to click the window segments in the Superframe viewer. Doesn't display the information yet, just need to write the html format for that and JS to get the relevant information and then make it appear like magic. Horay. Once the labelling of the superframe is done I may be able to move onto other things aside code for a bit.