Basic labelling and highlighting of WIndow segments in superframe
As title says, now highlighting which window the user has clicked on, and providing some of the GC Content data about it. I don't feel like it's useful enough yet, and the application as a whole feels lacking without the support of k-mer frequency analysis, but I'm running out of time and not sure how else I can talk about the GC Content vs ORF aside running artifical chimera data to see if any results come from it when looking at the visuals.