Further info on GC
When testing with an interesting mixed species file I found the chart visual show the obvious chimera, but the standard deviation obviously wouldn't pick this up.
Wrote a small bit of text to notify the user to look for this. Wondering if I can find these and highlight the difference in actual code for use in Superframe somehow, but running out of time to do my documentation so perhaps not.