Played with things, fixed a bug
Tried to get validation to display again but to no avail. Fixed a bug
with the gC windosize being wrong in the GC chart display though.
Somehow I was using a hard coded value for multiplication instead of the
actual GC Window size value. Not sure how that happened. Fixed it now
anyway.